+++ src/de/hhu/alobe/ba2016/mathe/KannStrahlenSchnitt.java
@@ -3,7 +3,7 @@ package de.hhu.alobe.ba2016.mathe;
 
 public interface KannStrahlenSchnitt {
 
-    float gibSchnittEntfernung(Strahl strahl);
+    double gibSchnittEntfernung(Strahl strahl);
 
     Vektor gibSchnittpunkt(Strahl strahl);
