+++ src/de/hhu/alobe/ba2016/physik/elemente/Auge/Augenlinse.java
@@ -6,7 +6,7 @@ import de.hhu.alobe.ba2016.physik.elemente.glasskoerper.Linse;
 
 public class Augenlinse extends Linse {
 
-    public Augenlinse(Auge auge, Vektor position, double brechzahl, float hoehe, float radius) {
+    public Augenlinse(Auge auge, Vektor position, double brechzahl, double hoehe, double radius) {
         super(auge.getOptischeBank(), position, brechzahl, hoehe, 0, radius, radius);
     }
